+++ apps/dispatch-service/src/sedml/sedml.service.ts
@@ -33,7 +33,6 @@ export class SedmlService {
 
   public async processSedml(id: string): Promise<void> {
     this.logger.log(`Processing SED-ML documents for simulation run '${id}'.`);
-    // get external url since combine service may not be local
     const url = this.endpoints.getRunDownloadEndpoint(false, id);
     const req = this.combine.getSedMlSpecs(undefined, url);
     const sedml = req.pipe(