+++ libs/config/common/src/lib/file-paths.ts
@@ -93,6 +93,11 @@ export class FilePaths {
     const dirPath = thumbnailType
       ? FilePaths.simulationRunThumbnailSubpath + '/' + thumbnailType
       : FilePaths.simulationRunContentsSubpath;
+
+    if (fileLocation?.startsWith('./')) {
+      fileLocation = fileLocation.substring(2);
+    }
+
     const filePath = fileLocation !== undefined ? `/${fileLocation}` : '';
     return this.getSimulationRunPath(runId, `${dirPath}${filePath}`);
   }